refactor(config): inline location reference plugin and refresh docs
